refactor(eval_station): migrate RoboDojo components to eval_station
- Update imports in debug_env_client.py and setup_policy_server.py to use eval_station instead of robodojo.
- Modify pyproject.toml to reflect the new package structure and dependencies.
- Revise README.md to align with the new eval_station setup.
- Introduce new files and modules for eval_station, including schemas, dispatch logic, and environment client functionality.
- Implement trial control and WebSocket adapter for the eval_station environment.

This refactor centralizes the evaluation station functionality under the eval_station package, improving organization and maintainability.
