+++ integrations/eval_station/dispatch/__init__.py
@@ -1,9 +1,9 @@
 """Dispatch orchestration: plan expansion, execution, and status."""
 
-from robodojo.dispatch.errors import normalize_execution_error
-from robodojo.dispatch.executor import notify_trial_failure, run_dispatch
-from robodojo.dispatch.planner import build_trial_runs, dispatch_for_trial
-from robodojo.dispatch.status import (
+from eval_station.dispatch.errors import normalize_execution_error
+from eval_station.dispatch.executor import notify_trial_failure, run_dispatch
+from eval_station.dispatch.planner import build_trial_runs, dispatch_for_trial
+from eval_station.dispatch.status import (
     STATUS_COMPLETED,
     STATUS_DONE,
     STATUS_FAILED,
