+++ integrations/eval_station/trial/__init__.py
@@ -1,6 +1,6 @@
 """Dispatch → deploy_cfg mapping helpers."""
 
-from robodojo.trial.config import (
+from eval_station.trial.config import (
     TrialRunConfig,
     build_trial_run_config,
     normalize_policy_name,
