+++ setup_policy_server.py
@@ -26,7 +26,7 @@ def main(deploy_cfg):
     model = model_class_func(deploy_cfg)
 
     if protocol == "robodojo_ws":
-        from robodojo.servers.policy_server import PolicyServer, PolicyServerConfig
+        from eval_station.servers.policy_server import PolicyServer, PolicyServerConfig
 
         server = PolicyServer(
             model,
