+++ utils/run_debug_env_client.sh
@@ -23,7 +23,7 @@ conda activate "${eval_env_conda_env}"
 echo -e "\033[34m[CLIENT] Activating Conda environment: ${eval_env_conda_env}\033[0m"
 echo -e "\033[34m[CLIENT] Connecting to server ${policy_server_ip}:${free_port}...\033[0m"
 
-PYTHONPATH="${root_dir}/XPolicyLab${PYTHONPATH:+:${PYTHONPATH}}"
+PYTHONPATH="${root_dir}/XPolicyLab/integrations:${root_dir}/XPolicyLab${PYTHONPATH:+:${PYTHONPATH}}"
 
 CLIENT_ARGS=(
     --dataset_name "${dataset_name}"
@@ -41,7 +41,7 @@ if [[ "${run_mode}" == "--run-once" ]]; then
     python "${root_dir}/XPolicyLab/debug_env_client.py" "${CLIENT_ARGS[@]}"
 else
     PYTHONWARNINGS=ignore::UserWarning \
-    python -m robodojo.servers.env_client_server \
+    python -m eval_station.servers.env_client_server \
         "${CLIENT_ARGS[@]}" \
         --serve-host 0.0.0.0 \
         --serve-port 19200
