+++ utils/run_real_policy_client.sh
@@ -36,7 +36,7 @@ conda activate "${eval_env_conda_env}"
 echo -e "\033[34m[CLIENT] Activating Conda environment: ${eval_env_conda_env}\033[0m"
 echo -e "\033[34m[CLIENT] Connecting to server ${policy_server_ip}:${free_port} (real env)...\033[0m"
 
-export PYTHONPATH="${root_dir}/src:${root_dir}/XPolicyLab:${root_dir}:${PYTHONPATH:-}"
+export PYTHONPATH="${root_dir}/src:${root_dir}/XPolicyLab/integrations:${root_dir}/XPolicyLab:${root_dir}:${PYTHONPATH:-}"
 
 CLIENT_ARGS=(
     --dataset_name "${dataset_name}"
@@ -62,7 +62,7 @@ if [[ -z "${action_type}" ]]; then
 fi
 
 PYTHONWARNINGS=ignore::UserWarning \
-python -m robodojo.servers.env_client_server \
+python -m eval_station.servers.env_client_server \
     "${CLIENT_ARGS[@]}" \
     --action-type "${action_type}" \
     --serve-host 0.0.0.0 \
